+++ src/attiny/ao_exti.c
@@ -33,7 +33,7 @@ ao_exti_setup_port(uint8_t pin, uint8_t mode, void (*callback)(void))
 {
 	(void) mode;
 	pcint_callback = callback;
-	pcint_mask = (1 << pin);
+	pcint_mask = (uint8_t) (1 << pin);
 	ao_exti_disable(PORTB, pin);
 	GIMSK |= (1 << PCIE);
 }
